Fix reload of all child components bug
Add trackBy function to *forLoop to to tell angular which elements are unique by the uuid key of the element. This fixes unnecessary re-render of the DOM and limits DOM updates to newly inserted or removed items. Fixes the issue of component "flash UI bug". 
In the reducer, the spread operator was only doing a shallow copy of the state object. Introducing elements in the child components to not update. By utilizing cloneDeep from lodash (to ensure immutability) fixes the issue.