Implement import survey functionality and option to import custom buttons to NgxBuilderViewerComponent component
NgxBuilderViewerComponent component accepts an option data as an Input(). It has the ability for users to import a survey and/or create custom buttons.
The import survey option takes a callback function returning  a survey from a get request. The ngx-survey library will then dispatch an event to the store to update the app state.
The custom buttons options have the following attributes: title, icon, text and callback. The callback returns the current survey state.